created vector_split.c file to hold logic for building command_array when splitting the command vector
