+++ src/execution/process_command.c
@@ -6,66 +6,15 @@
 /*   By: marvin <marvin@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/07/10 13:07:11 by jteissie          #+#    #+#             */
-/*   Updated: 2024/07/25 13:25:52 by marvin           ###   ########.fr       */
+/*   Updated: 2024/07/25 16:46:45 by marvin           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 
-char	*split_cmd_strings(t_vector *vector)
-{
-	size_t	vector_index;
-	size_t	word_index;
-	size_t	str_index;
-	size_t	len;
-	char	*string;
 
-	vector_index = 0;
-	len = 0;
-	str_index = 0;
-	while (!vector->buffer[vector_index])
-		vector_index++;
-	word_index = vector_index;
-	while (vector->buffer[word_index++])	
-		len++;
-	string = ft_calloc(len + 1, sizeof(char));
-	if (!string)
-		return (NULL);
-	word_index = vector_index;
-	while (vector->buffer[word_index])
-	{
-		string[str_index++] = vector->buffer[word_index];
-		vector->buffer[word_index] = '\0';
-		word_index++;
-	}
-	string[str_index] = '\0';
-	return (string);
-}
 
-char **make_command_array(t_vector *vector)
-{
-	size_t	words;
-	size_t	array_index;
-	char	**cmd_array;	
 
-	words = vector->word_count;
-	cmd_array = ft_calloc(words + 1, sizeof(char *));
-	if (!cmd_array)
-		return (NULL);
-	array_index = 0;
-	while (array_index < words)
-	{
-		cmd_array[array_index] = split_cmd_strings(vector);	
-		if (!cmd_array[array_index])
-		{
-			free_strarray(cmd_array);
-			return (NULL);
-		}
-		array_index++;
-	}
-	cmd_array[array_index] = NULL;
-	return (cmd_array);
-}
 
 void	execute_cmd(t_vector *cmd_vector, t_data *data)
 {
